move files from custom folder to apps folder as per suggestions
