+++ magentacloud25/apps/files_sharing/src/components/SharingPermissions.vue
@@ -264,12 +264,12 @@ import NcActionInput from '@nextcloud/vue/dist/Components/NcActionInput'
 import Tooltip from '@nextcloud/vue/dist/Directives/Tooltip'
 
 import SharesMixin from '../mixins/SharesMixin'
-import ShareTypes from '../../../../../../../nextcloud/apps/files_sharing/src/mixins/ShareTypes'
-import GeneratePassword from '../../../../../../../nextcloud/apps/files_sharing/src/utils/GeneratePassword'
+import ShareTypes from '../../../../../../nextcloud/apps/files_sharing/src/mixins/ShareTypes'
+import GeneratePassword from '../../../../../../nextcloud/apps/files_sharing/src/utils/GeneratePassword'
 import Vue from 'vue'
 import { mapGetters } from 'vuex'
-import ShareRequests from '../../../../../../../nextcloud/apps/files_sharing/src/mixins/ShareRequests'
-import ExternalShareAction from '../../../../../../../nextcloud/apps/files_sharing/src/components/ExternalShareAction'
+import ShareRequests from '../../../../../../nextcloud/apps/files_sharing/src/mixins/ShareRequests'
+import ExternalShareAction from '../../../../../../nextcloud/apps/files_sharing/src/components/ExternalShareAction'
 
 export default {
 	name: 'SharingPermissions',
