+++ magentacloud25/apps/files_sharing/src/files_sharing_tab.js
