+++ magentacloud25/apps/files_sharing/src/services/ConfigServiceCustom.js
