+++ magentacloud25/apps/files_sharing/src/share.js
